==> Programs/6.0 Descriptive_analysis.Rmd <==
---
title: "6.0 Descriptive chart"
author: "Phuc Quang Vu pqv9968"
date: "2024-01-13"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## R Markdown

```{r}
library(tidyr)
library(data.table)
#dir="/Users/quangphucvu/Desktop/ENAR Rerun/" 
setwd(paste0(dir,"Data"))
load("data_hyper_acc.rdata")
load("data_hyper_2011.rdata")
additional_cat_vars<-c("demo_age_cat","demo_race","bp_med_n_class","DPQ")
data_hyper_acc[additional_cat_vars]<-lapply(data_hyper_acc[additional_cat_vars],factor)
data_hyper_acc_design <- svydesign(id      = ~svy_id,
                          strata  = ~svy_strata,
                          weights = ~svy_weight_mec,
                          nest    = TRUE,
                          data    = data_hyper_acc)

nhanes_2013<-subset(data_hyper_acc_design,svy_year=="2013-2014")
nhanes_2015<-subset(data_hyper_acc_design,svy_year=="2015-2016")
nhanes_2017<-subset(data_hyper_acc_design,svy_year=="2017-2020")

nhanes_2011 <- svydesign(id     = ~svy_id,
                          strata  = ~svy_strata,
                          weights = ~svy_weight_mec,
                          nest    = TRUE,
                          data    = data_hyper_2011)



var_stats <- c("bp_control_accaha","demo_age_cat","demo_race","SIALANG","INDFMPIR","bp_med_n_class","bp_med_combination","bp_med_ace","bp_med_alpha","bp_med_beta","bp_med_ccb_dh","cc_ckd","DPQ","avg_DR1TMOIS","htn_aware")

svymean(~DPQ,nhanes_2013,na.rm=TRUE)
variables<-c()
mean_2011<-c()
SE_2011<-c()


mean_2013<-c()
SE_2013<-c()

mean_2015<-c()
SE_2015<-c()

mean_2017<-c()
SE_2017<-c()

for (var in var_stats){
  a<-svymean(~get(var),nhanes_2013,na.rm=TRUE)
  b<-svymean(~get(var),nhanes_2015,na.rm=TRUE)
  c<-svymean(~get(var),nhanes_2017,na.rm=TRUE)
  d<-svymean(~get(var),nhanes_2011,na.rm=TRUE)
  variables<-append(variables,var)
  if (is.factor(data_hyper_acc[[var]])){
    variables<-append(variables,names(a))
    mean_2011<-append(mean_2011,NA)
    mean_2011<-append(mean_2011,d)
    SE_2011<-append(SE_2011,NA)
    SE_2011<-append(SE_2011,SE(d))
    mean_2013<-append(mean_2013,NA)
    mean_2013<-append(mean_2013,a)
    SE_2013<-append(SE_2013,NA)
    SE_2013<-append(SE_2013,SE(a))
    mean_2015<-append(mean_2015,NA)
    mean_2015<-append(mean_2015,b)
    SE_2015<-append(SE_2015,NA)
    SE_2015<-append(SE_2015,SE(b))
    mean_2017<-append(mean_2017,NA)
    mean_2017<-append(mean_2017,c)
    SE_2017<-append(SE_2017,NA)
    SE_2017<-append(SE_2017,SE(c))
  }
  else {
  mean_2011<-append(mean_2011,d)
  SE_2011<-append(SE_2011,SE(d))
  mean_2013<-append(mean_2013,a)
  SE_2013<-append(SE_2013,SE(a))
  mean_2015<-append(mean_2015,b)
  SE_2015<-append(SE_2015,SE(b))
  mean_2017<-append(mean_2017,c)
  SE_2017<-append(SE_2017,SE(c))
  }
}

table_1<-data_frame(variables=variables,
                    mean_cycle_2011=mean_2011,
                    SE_cycle_2011=SE_2011,
                    mean_cycle_2013=mean_2013,
                    SE_cycle_2013=SE_2013,
                    mean_cycle_2015=mean_2015,
                    SE_cycle_2015=SE_2015,
                    mean_cycle_2017=mean_2017,
                    SE_cycle_2017=SE_2017
                    )

```

```{r}
table_1$variables<-gsub("*get\\(var\\)","",table_1$variables)
table_1
```

```{r}
mean<-table_1%>% select( mean_cycle_2011,mean_cycle_2013,mean_cycle_2015,mean_cycle_2017)
SE<-table_1%>% select( SE_cycle_2011,SE_cycle_2013,SE_cycle_2015,SE_cycle_2017)
colnames(mean)<-c("2011-2012","2013-2014","2015-2016","2017-2020")

# function to draw
bar<-function(index,y_col,scale,title){
  temp<-cbind(colnames(mean),data.frame(t(mean[index,]),row.names = NULL),data.frame(t(SE[index,]),row.names = NULL))
colnames(temp)<-c("Cycle",y_col,"se")
plot<-ggplot(temp) +
    geom_bar( aes(x=Cycle, y=scale*get(y_col)), stat="identity", fill="skyblue", alpha=0.7) +
  geom_text(aes(x=Cycle, y=scale*get(y_col),label = round(scale*get(y_col),2)), vjust = 5,size=5)+
    geom_errorbar( aes(x=Cycle, ymin=scale*(get(y_col)-se), ymax=scale*(get(y_col)+se)), width=0.4, colour="orange", alpha=0.9, size=1.5)+xlab("Cycle")+ylab(y_col)+ggtitle(title)+
  theme(plot.title = element_text(hjust = 0.5))+
  theme(text = element_text(size = 14),
        plot.title = element_text(size=14, face= "bold", colour= "black" ),
    axis.title.x = element_text(size=15, face="bold", colour = "black"),    
    axis.title.y = element_text(size=15, face="bold", colour = "black"))
return(plot)
}

stack<-function(range,title,order=NULL){
  temp<-table_1[range,]%>% select( variables,mean_cycle_2011,mean_cycle_2013,mean_cycle_2015,mean_cycle_2017)
colnames(temp)<-c("variables","2011-2012","2013-2014","2015-2016","2017-2020")

long_temp<-melt(temp, id.vars = c("variables"), variable.name = "Cycle")
if (!is.null(order)){
  long_temp$variables<-factor(long_temp$variables, levels=order, ordered=TRUE)
}



# Add text annotations to the plot
plot<-ggplot(long_temp, aes(fill=variables, y=100*value, x=Cycle)) + 
    geom_bar(position="stack", stat="identity")+
  geom_text(aes(label=ifelse(value>0.08,round(100*value,2),"") ), position = position_stack(vjust = 0.5),size=5)+
  ylab("Percentage")+ggtitle(title)+
  theme(plot.title = element_text(hjust = 0.5))+theme(text = element_text(size = 13))+
  theme(legend.title=element_blank(),
        plot.title = element_text(size=14, face= "bold", colour= "black" ),
    axis.title.x = element_text(size=14, face="bold", colour = "black"),    
    axis.title.y = element_text(size=14, face="bold", colour = "black"))+
  scale_fill_brewer(palette="Set3")+ 
  theme(legend.position="bottom")+
  guides(color = guide_legend(title.position = "top", 
                               hjust = 0.5, #centres the title horizontally
                              title.hjust = 0.5,
                              label.position = "bottom"),
         override.aes = list(size=0.1))+
  guides(fill=guide_legend(nrow =2))
  

return(plot)
}

```
```{r}

plot1<-bar(index=3,y_col="Percentage",scale=100,title="Prevalence of Blood Pressure control")
plot2<-stack(5:8,"Age Group",order=c("75+","65 to 74","45 to 64","18 to 44"))
plot3<-stack(10:14,"Race/Ethnicity")
table_1$variables[16:17]<-c("English","Spanish")
plot4<-stack(16:17,"Language")
table_1$variables[19:21]<-c("Middle","High","Low")
plot5<-stack(19:21,"Income",order = c("High","Middle","Low"))
plot6<-stack(23:27,"# of classes of antihypertensive medication",order=c("Four or more","Three","Two","One","None"))
plot7<-bar(index=30,y_col="Percentage",scale=100,title="Prevalence of using combination theraty")
plot8<-bar(index=33,y_col="Percentage",scale=100,title="Prevalence of using ACEis")
plot9<-bar(index=36,y_col="Percentage",scale=100,title="Prevalence of using Alpha blocker")
plot10<-bar(index=39,y_col="Percentage",scale=100,title="Prevalence of using Beta Blocker")
plot11<-bar(index=42,y_col="Percentage",scale=100,title="Prevalence of using CCB-DH")
plot12<-bar(index=45,y_col="Percentage",scale=100,title="Prevalence of CKD")
plot13<-stack(47:51,"Depresison",order=c("Severe","Moderately severe","Moderate","Mild","Minimal"))
plot14<-bar(index=52,y_col="mg",scale=1,title="Moisture")
plot15<-bar(index=55,y_col="Percentage",scale=100,title="Hypertension awareness")

```
```{r}

figure_1<-grid.arrange(plot1,plot15,plot6,plot8,plot12,plot9,ncol=2)

```
```{r}
ggsave(paste0(dir,"Tables and Figures/figure_1.png"), figure_1, height = 12, width = 8, dpi = 1080)

```
